requirements.txt
pandas
sqlalchemy
snowflake-connector-python
snowflake-snowpark-python
redshift-connector
boto3
openpyxl
